% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastclass.R
\docType{class}
\name{Blast-class}
\alias{Blast-class}
\alias{BlastInstantiate}
\title{S4 class}
\description{
An S4 representatio of blast objects. Please create using instances using \code{\link[=blast_parse]{blast_parse()}}.
}
\details{
Strong Hit = evalue < strong_hit_evalue & perc_identity > strong_hit_perc_identity & qcovs > strong_hit_query_coverage.

Required config file fields
percentage_identity,evalue_threshold,max_target_seqs,query,outfmt,num_seqs_blasted,max_hsps,subject_besthit,qcov_hsp_perc
}
\section{Slots}{

\describe{
\item{\code{blast_df}}{data.frame.}

\item{\code{blast_config}}{list. list of paramaters related to running of blast. This field is parsed from the blastn.config.tsv file produced by \code{\link[=blast_run]{blast_run()}}}

\item{\code{names_of_query_sequences}}{character. A vector containing the names of the sequences analysed using blast.}

\item{\code{strong_hit_evalue}}{numeric. The maximum threshold used to filter used to classify 'strong_hits'.}

\item{\code{strong_hit_perc_identity}}{numeric. The minimum threshold used to filter used to classify 'strong_hits'}

\item{\code{strong_hit_query_coverage}}{numeric. The minimum threshold used to filter used to classify 'strong_hits'}
}}

