% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastparse.R
\name{blast_parse}
\alias{blast_parse}
\title{Parse Blast Outputs}
\usage{
blast_parse(
  blast_path,
  blast_config_path = blast_path_to_config_path_guess(blast_path),
  blast_seqnames_path = blast_path_to_seqname_path_guess(blast_path),
  strong_hit_evalue = 1e-50,
  strong_hit_perc_identity = 97,
  strong_hit_query_coverage = 95,
  skip_sciname_mapping = FALSE
)
}
\arguments{
\item{blast_path}{path to tabular blast file (produced by \code{\link[=blast_run]{blast_run()}})}

\item{blast_config_path}{path to config file describing parameters used to run blastn}

\item{blast_seqnames_path}{path to a file listing the names of your query sequences. This file is autogenerated by \code{\link[=blast_run]{blast_run()}} and will be automatically found based on the blast_path unless the name has manually been changed}

\item{strong_hit_evalue, strong_hit_perc_identity, strong_hit_query_coverage, }{BLAST alignments are considered 'strong hits' if the alignment has an E-value > strong_hit_evalue, a percentage identity > strong_hit_perc_identity, and a query coverage of > strong_hit_query_coverage.}

\item{skip_sciname_mapping}{Don't attempt to map taxids to scientific names. This prevents downstream microbial analysis but can be useful if you just want to parse blast results generated using a local custom database lacking taxonomy information}
}
\value{
Blast object
}
\description{
Parses blast outputs produced by \code{\link[=blast_run]{blast_run()}}
}
