% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastanalyse.R
\name{blast_tabular_tophits_reactable}
\alias{blast_tabular_tophits_reactable}
\title{Blast Tabular tophits}
\usage{
blast_tabular_tophits_reactable(blast)
}
\arguments{
\item{blast}{Blast object (see \link{blast_parse})}
}
\value{
reactable
}
\description{
Blast Tabular tophits
}
