% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastanalyse.R
\name{blast_top_hits}
\alias{blast_top_hits}
\title{Find top hits
Takes a blast object and finds the most appopriate top hit for each.
Maybe this isn't the write way of looking at things. a lot of sequences might map
equally well to multiple sequences
Maybe just plot a barplot describing the number of sequences in which each species present as a tophit is present in the db
Find top hits}
\usage{
blast_top_hits(blast, add_unmapped_query_seqs = TRUE)
}
\arguments{
\item{blast}{Blast object (see \link{blast_parse})}
}
\value{
tibble with 1 row per query sequence
}
\description{
Takes a blast object and finds the absolute top hit for each.
Looks for minimum hit_rank (summed ranks of evalue, -perc_identity and -qcovs).
If ties exist will mark as 'multiple'
}
