% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastanalyse.R
\name{blast_top_hits_old}
\alias{blast_top_hits_old}
\title{Find top hits}
\usage{
blast_top_hits_old(blast, add_unmapped_query_seqs = TRUE)
}
\arguments{
\item{blast}{Blast object (see \link{blast_parse})}
}
\value{
tibble with 1 row per query sequence
}
\description{
Takes a blast object and finds the absolute top hit for each.
Looks for minimum hit_rank (summed ranks of evalue, -perc_identity and -qcovs).
Breaks ties based on which species has the greatest number of distinct subjects at the given hit_rank.
}
