% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastanalyse.R
\name{blast_top_microbes_barplot}
\alias{blast_top_microbes_barplot}
\title{Blast Top Microbes
Takes a blast object and finds the most appopriate top hit for each.
Maybe this isn't the write way of looking at things. a lot of sequences might map
equally well to multiple sequences
Maybe just plot a barplot describing the number of sequences in which each species present as a tophit is present in the db}
\usage{
blast_top_microbes_barplot(blast)
}
\arguments{
\item{blast}{blast object. see \code{\link[=blast_parse]{blast_parse()}}}
}
\value{
ggiraph plot
}
\description{
Blast Top Microbes
Takes a blast object and finds the most appopriate top hit for each.
Maybe this isn't the write way of looking at things. a lot of sequences might map
equally well to multiple sequences
Maybe just plot a barplot describing the number of sequences in which each species present as a tophit is present in the db
}
