% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastanalyse.R
\name{blast_top_microbes_donut}
\alias{blast_top_microbes_donut}
\title{Donut Pie of Most Common Microbes}
\usage{
blast_top_microbes_donut(blast)
}
\arguments{
\item{blast}{blast object. See \code{\link[=blast_parse]{blast_parse()}}}
}
\value{
ggiraph object
}
\description{
Donut Pie of Most Common Microbes
}
