% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastparse.R
\name{taxid2name}
\alias{taxid2name}
\title{NCBI Taxid to Scientific Name}
\usage{
taxid2name(taxids)
}
\arguments{
\item{taxids}{vector of taxids (numeric)}
}
\value{
scientific names for each taxid, or NA if no taxanomic match could be found
}
\description{
Faster taxid2name for long vectors with repeated taxids
}
