% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastparse.R
\name{taxid_to_higher_ranked_taxid}
\alias{taxid_to_higher_ranked_taxid}
\title{Taxid}
\usage{
taxid_to_higher_ranked_taxid(
  taxids,
  rank = c("species", "genus", "family", "order", "class", "clade", "phylum", "kingdom",
    "superkingdom")
)
}
\arguments{
\item{taxids}{ncbi taxonomy ID}

\item{rank}{what rank you're interested in (e.g. species, genus or family)}
}
\value{
first parent taxid at supplied rank or the input taxid if its already at this rank / higher  (string)
}
\description{
Takes a taxid and returns the parent taxid that matches the specified rank. If taxid is already that rank (or higher), will return the taxid unchanged
}
